% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_tmeans.R
\name{boot_tmeans}
\alias{boot_tmeans}
\title{Bootstrap distribution of trial means}
\usage{
boot_tmeans(
  formula = NULL,
  data,
  R = 500,
  var.names = c("y", "trial"),
  ci.type = "basic",
  level = 0.95,
  ncpus = 1
)
}
\arguments{
\item{formula}{standard R formula of the form response ~ trial}

\item{data}{data frame}

\item{R}{number of bootstrap samples}

\item{var.names}{alternative to formula}

\item{ci.type}{type of confidence interval (see boot::boot.ci)}

\item{level}{prediction level}

\item{ncpus}{number of cpus to use, passed to boot, negligible benefit for small datasets}
}
\value{
list with 'dat': data frame with the bootstrapped means, \cr 
         'boot.ci': confidence interval computed by 'boot.ci' \cr
         and 'pdi': prediction interval based on quantiles
}
\description{
Bootstrapped distribution of trial means
}
\examples{
\dontrun{
require(ggplot2)
data(soyrs)
## Simply calculate the trial means
tmns <- aggregate(lrr ~ Trial_ID, data = soyrs, FUN = mean)
## Bootstrapped stratified trial means
btm <- boot_tmeans(lrr ~ Trial_ID, data = soyrs, R = 2e3)

pdi.cf <- pred_int_conformal_df(formula = lrr ~ Trial_ID, x = soyrs)

btmd <- btm$dat
btm.q <- btm$pdi

ggplot() + xlab("lrr") + 
geom_density(data = btmd, aes(x = ys)) + 
geom_jitter(data = soyrs, aes(x = lrr, y = 2.5)) + 
geom_jitter(data = tmns, aes(x = lrr, y = 5), color = "blue", size = 1.2) + 
geom_point(aes(x = pdi.cf[1], y = -1), color = "orange", size = 1.2) + 
geom_errorbarh(mapping = aes(xmin = pdi.cf[2], xmax = pdi.cf[3],
                              y = -1), color = "orange", size = 1.2) +
geom_point(aes(x = btm.q[1], y = -2), color = "red", size = 1.2) + 
geom_errorbarh(aes(xmin = btm.q[2], xmax = btm.q[3], y = -2),
               color = "red", size = 1.2)

}
}
