% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_tmeans.R
\name{boot_tmeans2}
\alias{boot_tmeans2}
\title{hierarchical bootstrap}
\usage{
boot_tmeans2(
  formula = NULL,
  data,
  R = 500,
  var.names = c("y", "trial"),
  level = 0.95,
  ncpus = 1
)
}
\arguments{
\item{formula}{standard R formula, as in resp ~ trial}

\item{data}{data frame}

\item{R}{number of replicates}

\item{var.names}{optional alternative to formula}

\item{level}{probability level}

\item{ncpus}{number of cpus for parallel resampling (not implemented yet)}
}
\description{
This will implement hierarchical bootstrapping as oppossed to
stratified bootstrapping, which is the default
}
\examples{
\dontrun{
require(ggplot2)
## How does it compare to stratified resampling?
data(soyrs)
btm1 <- boot_tmeans(lrr ~ Trial_ID, data = soyrs)
btm2 <- boot_tmeans2(lrr ~ Trial_ID, data = soyrs)

dat <- data.frame(method = rep(c("boot","boot2"), each = nrow(btm2$dat)),
                  pos = rep(c(2,5), each = nrow(btm2$dat)),
                  ys = c(btm1$ys, btm2$dat$ys))
                  
ggplot(data = dat) + 
     xlab("lrr") +
     geom_point(aes(x = ys, y = pos, color = method)) + 
     geom_jitter(aes(x = ys, y = pos, color = method)) +
     geom_density(aes(x = ys, color = method))
     
## The two methods give nearly identical answers

}

}
