% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_pred_int.R
\name{cv_pred_int}
\alias{cv_pred_int}
\title{cv_pred_int}
\usage{
cv_pred_int(
  formula,
  data,
  method = c("tdist", "ntrial", "jags", "boot", "im"),
  level = 0.95,
  iter = 1,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{formula}{standard R formula, 'response ~ trial'}

\item{data}{a data frame}

\item{method}{either 'tdist', 'ntrial', 'jags', \sQuote{boot} or \sQuote{im}}

\item{level}{prediction level}

\item{iter}{number of iterations}

\item{quiet}{logical. FALSE by default}

\item{...}{additional arguments to be passed to the functions}
}
\description{
Test the coverage for different prediction methods using leave-one-out cross-validation
}
\examples{
\donttest{
 data(soyrs)
 t.tdist <- cv_pred_int(lrr ~ Trial_ID, data = soyrs)
}




}
