% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsum.R
\name{includes}
\alias{includes}
\title{Proportion of values in 'x' which are included in and interval}
\usage{
includes(x, lb, ub)
}
\arguments{
\item{x}{should be an object of class "numeric"}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\value{
a single value indicating the proportion of elements in 'x' which are contained in
a given interval [lb,ub]
}
\description{
Count how many elements of vector 'x' are included in a given interval.
}
\details{
if an element is contained in the interval [lb,ub] it will be counted.
https://www.varsitytutors.com/hotmath/hotmath_help/topics/interval-notation
}
\examples{
\dontrun{
x <- rnorm(100)
includes(x, -1.8, 1.8)
}
}
