% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_prob_pred_int.R
\name{max_prob_pred_int}
\alias{max_prob_pred_int}
\title{maximum probability for the prediciton interval 'small sample size'}
\usage{
max_prob_pred_int(x, n = 200, neval = 200, tol = 0.001,
  method = c("tdist", "conformal"), m.method = c("quantile",
  "deviation", "jackknife"))
}
\arguments{
\item{x}{should be a vector}

\item{n}{number of evaluations to perform in the line search of probabilities}

\item{neval}{number of evaluations in the conformal algorithm (see pred_int_conformal)}

\item{tol}{tolerance default is (0.001) see details}

\item{method}{either 'tdist' (assumes normality) or 'conformal' (distribution-free)}

\item{m.method}{method used to compute conformal prediction interval: either "quantile", "deviation" or "jackknife"}
}
\value{
a single value which represents a 'suggestion' for the maximum level of probability given the data
}
\description{
Calculates the maximum probability for a prediction interval
}
\details{
Calculate maximum probability for the prediction of a small sample size using conformal prediction

The idea is to find the maximum level of probability that will produce a conformal prediction interval
         which matches the minimum and maximum values in the observed sample. This is approximate and it depends on 
         the tolerance. The distance is calculated as abs(x.min - calc.lower.bound) + abs(x.max - calc.upper.bound).
         The deviation tolerance is calculated as a proportion of the range in the observed sample. i.e. (x.max - x.min)*tol
}
\note{
At this moment, with current testing, for method "tdist" this is not reliable and "opt_max_prob_pred_int" 
      should be used.
}
\examples{
\dontrun{
set.seed(12345)
mp.pdi <- max_prob_pred_int(rnorm(10), method = "conformal") 
## For this particular sample, the 'suggested' maximum is about 82\%
}


}
