% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_mcg.R
\name{mcmc.control}
\alias{mcmc.control}
\title{Control options for MCMCglmm within pred_int_mcg}
\usage{
mcmc.control(burnin = 5000, nitt = 50000, prior = NULL)
}
\arguments{
\item{burnin}{warmup or burnin iterations}

\item{nitt}{number of iterations for the MCMC method}

\item{prior}{prior specification as in the MCMCglmm package}
}
\description{
Control options for MCMCglmm within pred_int_mcg
}
