% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_aux.R
\name{npv}
\alias{npv}
\title{degrees of freedom calculation using 'emmeans'}
\usage{
npv(x, degfr = c("default", "zdist", "kr"), level = 0.95)
}
\arguments{
\item{x}{should be an object of class "lmerMod"}

\item{degfr}{degrees of freedom method ("default","zdist","kr")}

\item{level}{coverage level with default 0.95}
}
\value{
a single value 'numeric' to be used as input to a t-distribution
}
\description{
Calculates the nuisance parameter (degrees of freedom) using different methods
}
\details{
Calculate nuisance parameter (degrees of freedom)
}
\examples{
\dontrun{
## Include a meaningful example 
}


}
