% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_prob_pred_int.R
\name{opt_max_prob_pred_int}
\alias{opt_max_prob_pred_int}
\title{optimum maximum probability for the prediciton interval 'small sample size'}
\usage{
opt_max_prob_pred_int(x, method = c("tdist", "conformal", "npar"),
  m.method = c("quantile", "deviation", "jackknife"), interval = c(0,
  1), alpha.penalty = 0, scale = FALSE)
}
\arguments{
\item{x}{should be a vector}

\item{method}{either 'tdist' (assumes normality), 'conformal' (distribution-free), or non-parametric}

\item{m.method}{method used to compute conformal prediction interval: either "quantile", "deviation" or "jackknife"}

\item{interval}{maximum and minimum values for the optimization search}

\item{alpha.penalty}{whether to include a penalty for alpha}

\item{scale}{whether to scale the input vector}
}
\value{
a single value which represents a 'suggestion' for the maximum level of probability given the data
}
\description{
Finds the maximum probability for a prediction interval using optimization
}
\details{
Optimize the maximum probability for the prediction of a small sample size using either the 
t-distribution or conformal methods

The idea is to find the maximum level of probability that will produce a conformal prediction interval
         which matches the minimum and maximum values in the observed sample.  The distance is calculated as abs(x.min - calc.lower.bound) + abs(x.max - calc.upper.bound).
}
\examples{
\dontrun{
set.seed(12345)
x <- rnorm(10)
mp.pdi <- opt_max_prob_pred_int(x) 
## For this particular sample, the maximum is about 76\%
data(soyrs)
soyrs.s <- aggregate(lrr ~ Trial_ID, data = soyrs, FUN = mean)
mp.pdi.soy.tdist <- opt_max_prob_pred_int(soyrs.s$lrr)
mp.pdi.soy.conf <- opt_max_prob_pred_int(soyrs.s$lrr, method = "conformal")
print(round(c(mp.pdi.soy.tdist, mp.pdi.soy.conf1),2))
}


}
