% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_profile.R
\name{plot.pred_profile}
\alias{plot.pred_profile}
\title{plot.pred_profile}
\usage{
\method{plot}{pred_profile}(x, max.prob = TRUE, main = NULL, max.prob.col = "blue")
}
\arguments{
\item{max.prob}{logical whether to plot the optmiized maximum probability}

\item{main}{optional title for the plot}

\item{max.prob.col}{color for the plotting of the maximum probability}
}
\value{
profile plots
}
\description{
Plot Prediction Interval Profiles
}
