% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int.R
\name{pred_int}
\alias{pred_int}
\title{Prediction interval for different objects}
\usage{
pred_int(x, interval = c("prediction", "confidence"),
  method = c("tdist", "tdist2", "boot", "mcmc", "simulate", "predict",
  "metafor", "ntrial", "conformal"), m.method = "REML",
  degfr = c("default", "zdist", "kr"), level = 0.95, nsim = 500,
  var.names = NULL, formula = NULL, ...)
}
\arguments{
\item{x}{should be an object of class: "numeric", "data.frame", "lmerMod" or "MCMCglmm"}

\item{interval}{type of interval with default 'prediction'}

\item{method}{either 'tdist', 'tdist2', 'boot', 'mcmc', 'simulate', 'predict',
'metafor', 'ntrial' or 'conformal'.}

\item{m.method}{method for between-trial variance estimator (used in the 'metafor' package)}

\item{degfr}{degrees of freedom method default (n.k-2), zdist ("Inf") or "kr" (Kenward-Roger).
see package 'emmeans'}

\item{level}{coverage level with default 0.95}

\item{nsim}{number of simulations for the 'boot' method.}

\item{var.names}{variable names to be passed to the 'data.frame' methods}

\item{formula}{formula interface for 'data.frame' methods}

\item{...}{arguments to be passed to a few of the functions}
}
\value{
a prediction interval for a "new_trial"
}
\description{
Prediction interval for a variety of objects in the context of random-effects meta-analysis
}
\details{
This function will calculate a prediction interval in the context of
meta-analysis
As defined in Higgins et al. (2009) it considers the parameteric uncertainty
and the between study uncertainty, but there is more to this...
Implemented cases:

object = numeric \cr
object = lmerMod \cr
object = MCMCglmm \cr
object = data.frame \cr

For object = numeric the method is based on the 'lm' function

For object = 'lmerMod' there are three methods:
- tdist
- tdist2
- boot

For object = 'MCMCglmm' there are four methods:
- tdist
- mcmc
- simulate
- predict

There is also the 'ntrial' method, which needs to
be applied to a data.frame. It uses 'MCMCglmm'

Another method 'pred_int_metafor' uses the metafor pacakge
And it also requires a data.frame

the degrees of freedom argument (degfr) allows
for different methods other than n.k-2
Calculate prediction intervals for a variety of objects

TODO
}
\examples{
\dontrun{
## Using soybean row spacing
data(soyrs)
fit <- lmer(lrr ~ 1 + (1|Trial_ID), data = soyrs)
pdi <- pred_int(fit)
pdi 
}


}
