% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_tmeans.R
\name{pred_int_boot2}
\alias{pred_int_boot2}
\title{Bootstrap-based prediction intervals (method 2)}
\usage{
pred_int_boot2(formula, data, level = 0.95, R = 500, ncpus = 1)
}
\arguments{
\item{formula}{standard R formula of the form response ~ trial}

\item{data}{data frame}

\item{level}{prediction level}

\item{R}{number of bootstrap samples}

\item{ncpus}{number of cpus to use, passed to boot, negligible benefit for small datasets}
}
\value{
prediction interval
}
\description{
Bootstrap-based prediction intervals (method 2)
}
\details{
a simple wrapper for boot_tmeans2
}
