% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_aux.R
\name{pred_int_conformal}
\alias{pred_int_conformal}
\title{Prediction interval for a 'numeric' based on conformal inference}
\usage{
pred_int_conformal(x, neval = 200, level = 0.95,
  point.pred = c("median", "mean"))
}
\arguments{
\item{x}{should be an object of class "numeric"}

\item{neval}{number of evaluations to perform}

\item{level}{coverage level with default 0.95}

\item{point.pred}{point prediction either 'median' or 'mean'}
}
\value{
a prediction interval
}
\description{
Conformal inference prediction interval for a numeric
}
\details{
Calculate prediction interval based on conformal inference.

I wrote this function after reading this tutorial: https://cdsamii.github.io/cds-demos/conformal/conformal-tutorial.html
}
\examples{
\dontrun{
## Prediciton interval for simulated data
x <- rnorm(100)
pdi.t <- pred_int_tdist(x) 
pdi.c <- pred_int_conformal(x)
## Not really a prediction interval, but related
pdi.q <- quantile(x, probs = c(0.5, 0.025, 0.975))
}



}
