% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_im.R
\name{pred_int_im}
\alias{pred_int_im}
\title{Prediciton interval for Inferential Models}
\usage{
pred_int_im(x, level = 0.95, dist = c("z", "t"))
}
\arguments{
\item{x}{object of class \sQuote{merMod}}

\item{level}{probability level for interval coverage}
}
\value{
a prediction interval
}
\description{
Prediciton interval for Inferential Models
}
\examples{
\donttest{
require(lme4)
data(soyrs)
fit <- lmer(lrr ~ 1 + (1|Trial_ID), data = soyrs)
pdi <- pred_int(fit)
pdi 
## New formula
pdi2 <- pred_int_im(fit)

}
}
