% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_jags.R
\name{pred_int_jags}
\alias{pred_int_jags}
\title{pred_int_jags}
\usage{
pred_int_jags(formula, data, df = 1, level = 0.95, quiet = FALSE,
  return = c("pdi", "all"))
}
\arguments{
\item{formula}{standard R formula response ~ trial}

\item{data}{data frame}

\item{df}{degrees of freedom for the t-distribution. If 'NA' it will be treated as a nuisance parameter}

\item{level}{probability level}

\item{quiet}{whether to display progress of sampling}

\item{return}{type of information to return either 'pdi' just the prediction interval or 'all' which include the output from the model and Gelman diagnostics}
}
\value{
prediction interval (or more detail)
}
\description{
Prediction intervals using JAGS
}
