% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_jb.R
\name{pred_int_jb}
\alias{pred_int_jb}
\title{Prediction Interval for Random-effects Meta-analysis, which uses a combination of Jackknife and Bootstrap}
\usage{
pred_int_jb(formula, data, level = 0.95, iter = 10, R = 50,
  var.names = NULL)
}
\arguments{
\item{formula}{standard formula with response ~ trial}

\item{data}{should be an object of class "data.frame"}

\item{level}{coverage level with default 0.95}

\item{iter}{number of iterations for the bootstrap level 1}

\item{R}{number of iterations for the bootstrap level 2}

\item{var.names}{variable names to be passed to the 'data.frame' methods}
}
\value{
a list with 'pdi' prediction intervals, 'cf' correction factor and 'coverage' empirical coverage
}
\description{
Prediction for Random-effects Model using Subsampling
}
\examples{
\dontrun{
## Using soybean row spacing
data(soyrs)
pdi.rs <- pred_int_jb(formula = lrr ~ Trial_ID, data = soyrs) 
data(soysff)
pdi.ff <- pred_int_jb(formula = lrr ~ Trial_ID, data = soysff) 
}

}
