% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_mcg.R
\name{pred_int_mcg_ntrial}
\alias{pred_int_mcg_ntrial}
\title{Prediction interval for a data.frame using package 'MCMCglmm' and method 'ntrial'}
\usage{
pred_int_mcg_ntrial(x, var.names = c("y", "trial"), level = 0.95)
}
\arguments{
\item{x}{should be an object of class "data.frame"}

\item{var.names}{variable names for 'response' and 'id': c("y","trial")}

\item{level}{coverage level with default 0.95}
}
\value{
a prediction interval for a "new_trial"
}
\description{
Prediction interval for a data.frame using package 'MCMCglmm' and method 'ntrial'
}
\details{
Calculate prediction intervals using the 'MCMCglmm' pacakge and method 'ntrial'

see function 'MCMCglmm' in package 'MCMCglmm'. The method adds an extra row to 
the data.frame with missing response and then calculates the 'prediciton' for that new row
using the quantiles from the posterior distribution.
}
\examples{
\dontrun{
## Using soybean row spacing
data(soyrs)
pdi <- pred_int_mcg_ntrial(soyrs, var.names = c("lrr","Trial_ID"))
pdi 
}


}
