% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_mcg.R
\name{pred_int_mcg_tdist}
\alias{pred_int_mcg_tdist}
\title{Prediction interval for an object of class 'MCMCglmm' using method 'tdist'}
\usage{
pred_int_mcg_tdist(x, level = 0.95)
}
\arguments{
\item{x}{should be an object of class 'MCMCglmm'}

\item{level}{coverage level with default 0.95}
}
\value{
a prediction interval
}
\description{
Prediction interval for an object of class 'MCMCglmm' using method 'tdist'
}
\details{
see function 'MCMCglmm' in package 'MCMCglmm'. This method extracts the point estimates
for the mean, between-trial variance and within trial variance and plugs them in the 'tdist'
prediction formula
}
