% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_metafor.R
\name{pred_int_metafor}
\alias{pred_int_metafor}
\title{Prediction interval for a data.frame assuming effect size 'ROM' and 'metafor' package}
\usage{
pred_int_metafor(
  x,
  method = "REML",
  interval = c("prediction", "confidence"),
  var.names = c("trt", "ctrl", "ID"),
  level = 0.95
)
}
\arguments{
\item{x}{should be an object of class "data.frame"}

\item{method}{method used in computing the between-trial variance (default = "REML")}

\item{interval}{type of interval with default being "prediction"}

\item{var.names}{variable names for 'treatment', 'control' and 'id': c("trt","ctrl","ID")}

\item{level}{coverage level with default 0.95}
}
\value{
list with elements 'pdi' (prediction interval) and 'I2' (heterogeneity ratio)
}
\description{
Prediction interval for a data.frame assuming effect size 'ROM'
}
\details{
Calculate prediction intervals using the 'metafor' pacakge

see functions in metafor package 'escalc', 'rma' and 'predict.rma'
}
\examples{
\dontrun{
## Using soybean row spacing
data(soyrs)
pdi.I2 <- pred_int_metafor(soyrs, var.names = c("TRT1_Yld","TRT2_Yld","Trial_ID")) 
}


}
