% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_tdist.R
\name{pred_int_tdist}
\alias{pred_int_tdist}
\title{Prediction interval for a 'numeric' based on the t-distribution}
\usage{
pred_int_tdist(x, level = 0.95, interval = c("prediction",
  "confidence"))
}
\arguments{
\item{x}{should be an object of class "numeric"}

\item{level}{coverage level with default 0.95}

\item{interval}{type of interval, by default is "prediction"}
}
\value{
a prediction interval
}
\description{
Simple wrapper for 'lm' and 'predict' which returns a prediction interval. 
This method assumes iid normally distributed observations.
}
\details{
Calculate prediction interval based on the t-distribution. It uses 
'lm' and 'predict'

the prediction is at the mean of 'x'. Missing values are not allowed.
}
\examples{
\dontrun{
## Prediciton interval for simulated data
x <- rnorm(100)
pdi <- pred_int_tdist(x) 
}



}
\seealso{
function 'lm' and 'predict.lm'
}
