% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_tdist.R
\name{pred_int_tdist_df}
\alias{pred_int_tdist_df}
\title{Prediction interval for a 'data.frame' based on the t-distribution}
\usage{
pred_int_tdist_df(x, level = 0.95, neval = 500,
  interval = c("prediction", "confidence"), var.names = c("y",
  "trial"), formula = NULL)
}
\arguments{
\item{x}{should be an object of class "data.frame"}

\item{level}{coverage level with default 0.95}

\item{neval}{number of evaluations (default = 500)}

\item{interval}{type of interval, by default is "prediction"}

\item{var.names}{names of variables for response and 'trial'}

\item{formula}{alternative formula interface}
}
\value{
a prediction interval
}
\description{
Simple wrapper which 'bootstraps' the trial means.
}
\details{
Calculate prediction interval based on the t-distribution. It uses 
'pred_int_tdist_df'

the prediction is at the mean of 'x'. Missing values are not allowed.
}
\examples{
\dontrun{
## Prediciton interval for soybean data
data(soyrs)
pdi <- pred_int_tdist_df(soyrs, var.names = c("lrr","Trial_ID")) 
}

}
\seealso{
function 'pred_int_tdist_df'
}
