% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{soysff}
\alias{soysff}
\title{Soybean Stratego Foliar Fungicide}
\format{A data frame with 200 rows and 6 variables:
\describe{
  \item{Year}{ -integer- year that the trial was conducted.}
  \item{Trial_ID}{ -factor- given name for the specific trial.}
  \item{Rep}{ -integer- rep number.}
  \item{TRT_Yld}{ -numeric- yield for treated.}
  \item{CTR_Yld}{ -numeric- yield for untreated (control).}
  \item{lrr}{ -numeric- logarithm of the 'response ratio' i.e. log(TRT_Yld/CTR_Yld).}
}}
\source{
\url{https://analytics.iasoybeans.com/cool-apps/ISOFAST/}
}
\usage{
soysff
}
\description{
A dataset containing the Year, trial id, rep, yield of treatment, yield of control and the
logarithm of the response ratio
}
\keyword{datasets}
