% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pred_int.R
\name{test_pred_int}
\alias{test_pred_int}
\title{test_pred_int}
\usage{
test_pred_int(
  formula,
  data,
  method = c("tdist", "ntrial", "jags", "boot"),
  level = 0.95,
  iter = 1,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{formula}{standard R formula, 'response ~ trial'}

\item{data}{a data frame}

\item{method}{either 'tdist', 'ntrial', 'jags' or 'boot'}

\item{level}{prediction level}

\item{iter}{number of iterations}

\item{quiet}{logical. FALSE by default}

\item{...}{additional arguments to be passed to the functions}
}
\description{
Test the coverage for different prediction methods
}
\examples{
\dontrun{
 data(soyrs)
 t.tdist <- test_pred_int(lrr ~ Trial_ID, data = soyrs)
}




}
