% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsum.R
\name{tsum}
\alias{tsum}
\title{Summarize a data.frame according to a given 'trial'}
\usage{
tsum(
  x,
  var.names = c("y", "trial"),
  method = c("ici", "lm", "min-max"),
  level = 0.95,
  order = TRUE,
  add.id = TRUE,
  formula = NULL
)
}
\arguments{
\item{x}{should be an object of class "data.frame"}

\item{var.names}{names of variables 'y' (response) and 'trial' variable to summarize by.}

\item{method}{'ici' individual confidence interval, 'lm' model-based confidence interval 
(with typical assumptions of a linear model), 'min-max' minimum and maximum values}

\item{level}{coverage level with default 0.95}

\item{order}{whether to order the summary by mean of response}

\item{add.id}{whether to add the 'id' label}

\item{formula}{alternative formula interface, if this is used 'var.names' is ignored.}
}
\value{
a data.frame with the indicated summaries
}
\description{
Summarize data according to a 'trial' variable using either individual
 confidence intervals, model-based confidence intervals ('lm') or 'min-max'
}
\details{
Custom function to summarize trial data

this function is for somewhat of narrow use case, but similar things can be accomplished
using 'tidyverse' type operations. It uses 'aggregate'.
}
\examples{
\dontrun{
## Summarize trial data
data(soyrs)
soyrs.s <- tsum(soyrs, var.names = c("lrr","Trial_ID")) 

ggplot(data = soyrs.s, aes(x = m, y = id)) + 
geom_point() + xlab("log RR") + ylab("ID") +
geom_vline(xintercept = 0)

}
}
