% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int.R
\name{pred_int}
\alias{pred_int}
\title{Prediction interval for different objects}
\usage{
pred_int(
  x,
  interval = c("prediction", "confidence"),
  method = c("tdist", "tdist2", "boot", "mcmc", "simulate", "predict", "metafor",
    "ntrial", "conformal"),
  m.method = c("REML", "DL", "HE", "SJ", "ML", "REML", "EB", "HS", "GENQ"),
  c.method = c("quantile", "deviation", "jackknife"),
  degfr = c("default", "zdist", "kr"),
  level = 0.95,
  nsim = 500,
  var.names = NULL,
  formula = NULL,
  ...
)
}
\arguments{
\item{x}{should be an object of class: "numeric", "data.frame", "lmerMod" or "MCMCglmm"}

\item{interval}{type of interval with default 'prediction'}

\item{method}{either 'tdist', 'tdist2', 'boot', 'mcmc', 'simulate', 'predict',
'metafor', 'ntrial' or 'conformal'.}

\item{m.method}{method for between-trial variance estimator (used in the 'metafor' package)}

\item{c.method}{method for conformal prediction. See ?pred_int_conformal}

\item{degfr}{degrees of freedom method default (n.k-2), zdist ("Inf") or "kr" (Kenward-Roger).
see package 'emmeans'}

\item{level}{coverage level with default 0.95}

\item{nsim}{number of simulations for the 'boot' method.}

\item{var.names}{variable names to be passed to the 'data.frame' methods}

\item{formula}{formula interface for 'data.frame' methods}

\item{...}{arguments to be passed to a few of the functions}
}
\value{
a prediction interval for a "new_trial"
}
\description{
Prediction interval for a variety of objects in the context of random-effects meta-analysis
}
\details{
This function will calculate a prediction interval in the context of
meta-analysis
As defined in Higgins et al. (2009) it considers the parameteric uncertainty
and the between study uncertainty, but there is more to this...
Implemented cases:

object of class 'numeric' \cr
object of class 'lmerMod' \cr
object of class 'MCMCglmm' \cr
object of class 'data.frame' \cr

For object of class numeric method t-dist is based on the 'lm' function and predict
There is also a method based on conformal prediction. See 'pred_int_conformal'

For object of class 'lmerMod' there are three methods: \cr
- tdist \cr
- tdist2 \cr
- boot (uses lme4::bootMer) \cr

For object of class 'MCMCglmm' there are four methods: \cr
- tdist \cr
- mcmc \cr
- simulate \cr
- predict \cr

There is also the 'ntrial' method, which needs to
be applied to a data.frame. It uses 'MCMCglmm'

Another method 'pred_int_metafor' uses the metafor pacakge
And it also requires a data.frame

the degrees of freedom argument (degfr) allows
for different methods other than n.k-2
Calculate prediction intervals for a variety of objects

The main reference to understand the rationale behind this function is \cr
Higgins et al. (2009) A re-evaluation of random-effect meta-analysis
}
\examples{
\dontrun{
## Using soybean row spacing
data(soyrs)
fit <- lmer(lrr ~ 1 + (1|Trial_ID), data = soyrs)
pdi <- pred_int(fit)
pdi 
}


}
