% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_int_mcg.R
\name{pred_int_mcg_ntrial}
\alias{pred_int_mcg_ntrial}
\title{Prediction interval for a data.frame using package 'MCMCglmm' and method 'ntrial'}
\usage{
pred_int_mcg_ntrial(formula, data, level = 0.95)
}
\arguments{
\item{formula}{an object of class formula with response ~ group}

\item{data}{a data frame which contains the response and the grouping}

\item{level}{coverage level with default 0.95}
}
\value{
a prediction interval for a "new_trial"
}
\description{
Prediction interval for a data.frame using package 'MCMCglmm' and method 'ntrial'
}
\details{
Calculate prediction intervals using the 'MCMCglmm' pacakge and method 'ntrial'

see function 'MCMCglmm' in package 'MCMCglmm'. The method adds an extra row to 
the data.frame with missing response and then calculates the 'prediciton' for that new row
using the quantiles from the posterior distribution.
}
\examples{
\dontrun{
## Using soybean row spacing
data(soyrs)
pdi <- pred_int_mcg_ntrial(lrr ~ Trial_ID, data = soyrs)
pdi 
}


}
